update and address feedback
