+++ protocol-designer/src/components/organisms/StepSummary/StyledTrans.tsx
@@ -1,6 +1,4 @@
 import { Trans, useTranslation } from 'react-i18next'
-import first from 'lodash/first'
-import last from 'lodash/last'
 
 import {
   ALIGN_CENTER,
@@ -11,23 +9,6 @@ import {
   Tag,
   WRAP,
 } from '@opentrons/components'
-
-export const getWellsForStepSummary = (
-  targetWells: string[],
-  labwareWells: string[]
-): string => {
-  if (targetWells.length === 1) {
-    return targetWells[0]
-  }
-  const firstElementIndexOffset = labwareWells.indexOf(targetWells[0])
-  const isInOrder = targetWells.every(
-    (targetWell, i) =>
-      labwareWells.indexOf(targetWell) === firstElementIndexOffset + i
-  )
-  return isInOrder
-    ? `${first(targetWells)}-${last(targetWells)}`
-    : `${targetWells.length} wells`
-}
 
 interface StyledTransProps {
   i18nKey: string
