+++ protocol-designer/src/components/organisms/StepSummary/utils.ts
@@ -1,0 +1,19 @@
+import first from 'lodash/first'
+import last from 'lodash/last'
+
+export const getWellsForStepSummary = (
+  targetWells: string[],
+  labwareWells: string[]
+): string => {
+  if (targetWells.length === 1) {
+    return targetWells[0]
+  }
+  const firstElementIndexOffset = labwareWells.indexOf(targetWells[0])
+  const isInOrder = targetWells.every(
+    (targetWell, i) =>
+      labwareWells.indexOf(targetWell) === firstElementIndexOffset + i
+  )
+  return isInOrder
+    ? `${first(targetWells)}-${last(targetWells)}`
+    : `${targetWells.length} wells`
+}
